renamed expansion to extension
