+++ src/main/kotlin/internals/mappers/NameMapper.kt
@@ -1,9 +1,9 @@
 package internals.mappers
 
-import internals.expansions.camelCase
-import internals.expansions.lowerSnakeCase
-import internals.expansions.upperCamelCase
-import internals.expansions.upperSnakeCase
+import internals.extensions.camelCase
+import internals.extensions.lowerSnakeCase
+import internals.extensions.upperCamelCase
+import internals.extensions.upperSnakeCase
 
 @Suppress("unused")
 object NameMapper {
